refactor: Move SearchFilesViewModel to file
